requirements.txt
openai>=1.30.0
Flask==3.0.3
python-dotenv
numpy
python-frontmatter
